requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
earthengine-api==0.1.396
google-auth==2.29.0
python-dotenv
requests
